+++ webviz_subsurface/plugins/_surface_with_seismic_cross_section.py
@@ -15,8 +15,9 @@ from webviz_config import WebvizSettings
 from webviz_config.webviz_store import webvizstore
 from webviz_config.utils import calculate_slider_step
 
+from webviz_subsurface._models import SurfaceLeafletModel
 from .._datainput.seismic import load_cube_data
-from .._datainput.surface import make_surface_layer, get_surface_fence
+from .._datainput.surface import get_surface_fence
 
 
 class SurfaceWithSeismicCrossSection(WebvizPluginABC):
@@ -233,12 +234,7 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                                     autoScaleMap=True,
                                     minZoom=-5,
                                     updateMode="update",
-                                    # draw_toolbar_polyline=True,
-                                    layers=[
-                                        # make_surface_layer(
-                                        # xtgeo.RegularSurface(), name="surface"
-                                        # )
-                                    ],
+                                    layers=[],
                                     drawTools={
                                         "drawMarker": False,
                                         "drawPolygon": False,
@@ -247,38 +243,11 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                                     },
                                     mouseCoords={"position": "bottomright"},
                                     colorBar={"position": "bottomleft"},
-                                ),
-                                html.Div(
-                                    children=[
-                                        dcc.RadioItems(
-                                            id=self.uuid("hillshade"),
-                                            labelStyle={
-                                                "display": "inline-block",
-                                                "text-align": "justify",
-                                            },
-                                            options=[
-                                                {
-                                                    "value": None,
-                                                    "label": "No hillshading",
-                                                },
-                                                {
-                                                    "value": "soft-hillshading",
-                                                    "label": "Soft hillshading",
-                                                },
-                                                {
-                                                    "value": "hillshading",
-                                                    "label": "Hillshading",
-                                                },
-                                                {
-                                                    "value": "hillshading_shadows",
-                                                    "label": "Hillshading with shadows",
-                                                },
-                                            ],
-                                            value=None,
-                                            persistence=True,
-                                            persistence_type="session",
-                                        ),
-                                    ]
+                                    switch={
+                                        "value": False,
+                                        "disabled": False,
+                                        "label": "Hillshading",
+                                    },
                                 ),
                             ],
                         )
@@ -396,7 +365,7 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                 Input(self.ids("cube"), "value"),
                 Input(self.ids("color-values"), "value"),
                 Input(self.ids("color-scale"), "colorscale"),
-                Input(self.uuid("hillshade"), "value"),
+                Input(self.ids("map-view"), "switch"),
             ],
         )
         def _render_surface(
@@ -413,16 +382,15 @@ e.g. [xtgeo](https://xtgeo.readthedocs.io/en/latest/).
                 color = colorscale if colorscale else color
                 cube = load_cube_data(get_path(cubepath))
                 surface.slice_cube(cube)
-
-            s_layer = make_surface_layer(
-                surface,
-                name="surface",
-                min_val=min_val,
-                max_val=max_val,
-                color=color,
-                shader_type=hillshade,
-            )
-            return [s_layer]
+            return [
+                SurfaceLeafletModel(
+                    surface,
+                    name="surface",
+                    clip_min=min_val,
+                    clip_max=max_val,
+                    apply_shading=hillshade.get("value", False),
+                ).layer
+            ]
 
         @app.callback(
             Output(self.ids("fence-view"), "figure"),